+++ tests/mst/__init__.py
@@ -1,0 +1,1 @@
+"""MST-level tests."""
